+++ integrations/coinbase/src/main/java/org/dash/wallet/integrations/coinbase/model/CoinBaseAccountAddressResponse.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2021 Dash Core Group.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package org.dash.wallet.integrations.coinbase.model
 
 import android.os.Parcelable
@@ -93,4 +109,16 @@ data class DataItem(
 
     @field:SerializedName("id")
     val id: String? = null
+) : Parcelable
+
+@Parcelize
+data class Pagination(
+    val ending_before: String? = null,
+    val limit: Int? = null,
+    val next_starting_after: String? = null,
+    val next_uri: String? = null,
+    val order: String? = null,
+    val previous_ending_before: String? = null,
+    val previous_uri: String? = null,
+    val starting_after: String? = null
 ) : Parcelable
